requirements.txt
pyrogram>=2.0.106
TgCrypto>=1.2.5
pypdf>=4.0.0
python-dotenv>=1.0.0
Flask>=2.0.0
psutil>=5.9.0
requests>=2.31.0